Get rid of sepaeate processor manager
